update home structure
